+++ main.js
@@ -249,11 +249,12 @@ define(function (require, exports, module) {
     global_prefs.on('change', 'fontFamily', applyFontChanges);
 
     function applyFontChanges() {
-        $('.console-element')
+        $('.console-element, .link_to_diff.console-element')
             .css('font-size', global_prefs.get('fontSize'))
             .css('font-family', global_prefs.get('fontFamily'));
 
-        $('.brackets-nodejs-integration-debugger-log')
-            .css('font-size', global_prefs.get('fontSize'));
+        $('.brackets-nodejs-integration-debugger-log, .brackets-nodejs-integration-debugger-log span.var-name, .brackets-nodejs-integration-debugger-log span.var-value')
+            .css('font-size', global_prefs.get('fontSize'))
+            .css('font-family', global_prefs.get('fontFamily'));
     }
 });